// types/db.ts
export type CartItem = {
  product_id: number
  name: string
  quantity: number
  price: number
  image: string
  is_bundle?: boolean
}